% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/durnit.R
\name{durnit}
\alias{durnit}
\title{Knit all .Rmd files in a directory.}
\usage{
durnit(input_dir, output_dir)
}
\arguments{
\item{input_dir}{Input directory}

\item{output_dir}{Output directory}
}
\value{
Knitted documents in the appropriate output format
}
\description{
Knit all .Rmd files in a directory.
}
